using the DB No more files for graph metadata and vertices metadata. This information is saved in the Db and retrived when needed.
Rempved the DB table id_to_eth and merged it into the graph_vertex_table. This is more convenient. This makes sense as the mapping between ID and Eth_address is specific to a single graph vertex, therefore it makes sense for the mapping to exists in the graph_vertex table.