renamed galleryTypes to gallery Categories.
Custom admin view for graphCategory that checks before deleting a categoty. Custom admin view fro graph that shows also foregin key.

Removed redundant services to infer the vertex & edge table name: using method in configuration.
Vertex and edge table names do not depend on the graph's name.

Graph category is provided as a string when running GTM: we need to resolve it to it's ID with the DB, this is done in gtm command