fix(account & metadata)
If account_type and metadata tables are empty
the code currently breaks (None is not handled).
As a temporary fix I added some fake data in the
script that sets up the DB. In the future we need to handle when a graph
has not matches in the acount and metadata table

In the client code we now at least check if the code is None
but this is not enough.