refactor(Edge ORM, columns)
Before the Edge was not using the ORM (probably
because it was complicated).
But actually it can and should (less code to maintain).

Modified DB setup script because changed column names of Edge table.

Added backref to Vertex in the Edge model.

All Edge methods related to obtain a limited amoun of random edges
from a vertex (because there can be many many edges) are not needed.
This is both because I use ORM and because I changed the randomisation
method to an OFFSET/LIMIT based approach.

The conversion from WKB to list of float is done on
the schema (Marshmallow)