fix(consolidate labels & acc_types)
Before vertex labels and account types used to be in 2  separate DB
tables, now they are on the same.

This means that instead of 2 CSV files there is only one to be provided
to the setup script.

The creation of the vertex_metadata table happens in the script that
loads the CSV  file with the COPY command, this is because the COPY
command needs an empty table when executed

Also GTM code that queries the metadata of the vertices being created
has been modifies to accomodate the new server endpoint